+++ driverlib/f28003x/driverlib/epg.h
@@ -51,6 +51,7 @@ extern "C" {
 #include "inc/hw_types.h"
 #include "inc/hw_epg.h"
 #include "debug.h"
+#include "cpu.h"
 
 //*****************************************************************************
 //
@@ -415,10 +416,11 @@ EPG_selectEPGOutput(uint32_t base, EPG_OUT gPinNum,
     uint32_t regValue, bitFieldLoc, bitFieldVal;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(gPinNum <= 7U);
+    ASSERT((uint16_t)gPinNum <= 7U);
 
-    bitFieldLoc = 1U << (gPinNum + EPG_GCTL0_EPGOUTSEL_S);
-    bitFieldVal = (uint32_t)gPinOutSel << (gPinNum + EPG_GCTL0_EPGOUTSEL_S);
+    bitFieldLoc = 1UL << ((uint16_t)gPinNum + EPG_GCTL0_EPGOUTSEL_S);
+    bitFieldVal = (uint32_t)gPinOutSel << ((uint16_t)gPinNum +
+                                           EPG_GCTL0_EPGOUTSEL_S);
 
     regValue = (HWREG(base + EPG_O_GCTL0) & (~bitFieldLoc));
     regValue |= bitFieldVal;
@@ -453,13 +455,13 @@ EPG_selectSignalOutput(uint32_t base, EPG_OUT gPinNum,
     uint32_t regValue, bitFieldLoc, bitFieldVal;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(gPinNum <= 7U);
+    ASSERT((uint16_t)gPinNum <= 7U);
 
-    bitFieldLoc = EPG_GCTL3_EPGOUT0_SIGOUTSEL_M <<
-                  (gPinNum * EPG_GCTL3_EPGOUT_SIGOUTSEL_BIT_L);
+    bitFieldLoc = (uint32_t)EPG_GCTL3_EPGOUT0_SIGOUTSEL_M <<
+                  ((uint32_t)gPinNum * EPG_GCTL3_EPGOUT_SIGOUTSEL_BIT_L);
 
     bitFieldVal = (uint32_t)sigGenOutput <<
-                  (gPinNum * EPG_GCTL3_EPGOUT_SIGOUTSEL_BIT_L);
+                  ((uint32_t)gPinNum * EPG_GCTL3_EPGOUT_SIGOUTSEL_BIT_L);
 
     regValue = HWREG(base + EPG_O_GCTL3) & ~bitFieldLoc;
     regValue |= bitFieldVal;
@@ -495,13 +497,13 @@ EPG_selectClkOutput(uint32_t base, EPG_OUT gPinNum,
     uint32_t regValue, bitFieldLoc, bitFieldVal;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(gPinNum <= 7U);
+    ASSERT((uint16_t)gPinNum <= 7U);
 
-    bitFieldLoc = EPG_GCTL2_EPGOUT0_CLKOUTSEL_M <<
-                  (gPinNum * EPG_GCTL2_EPGOUT_CLKOUTSEL_BIT_L);
+    bitFieldLoc = (uint32_t)EPG_GCTL2_EPGOUT0_CLKOUTSEL_M <<
+                  ((uint32_t)gPinNum * EPG_GCTL2_EPGOUT_CLKOUTSEL_BIT_L);
 
     bitFieldVal = (uint32_t)clkGenOutput <<
-                  (gPinNum * EPG_GCTL2_EPGOUT_CLKOUTSEL_BIT_L);
+                  ((uint32_t)gPinNum * EPG_GCTL2_EPGOUT_CLKOUTSEL_BIT_L);
 
     regValue = HWREG(base + EPG_O_GCTL2) & ~bitFieldLoc;
     regValue |= bitFieldVal;
@@ -538,13 +540,13 @@ EPG_selectSigGenClkSource(uint32_t base, EPG_SIGGEN sigGenNum,
     uint32_t regValue, bitFieldLoc, bitFieldVal;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint16_t)sigGenNum <= 0U);
 
-    bitFieldLoc = EPG_GCTL1_SIGGEN0_CLKSEL_M <<
-                  (sigGenNum * EPG_GCTL1_SIGGEN_CLKSEL_BIT_L);
+    bitFieldLoc = (uint32_t)EPG_GCTL1_SIGGEN0_CLKSEL_M <<
+                  ((uint32_t)sigGenNum * EPG_GCTL1_SIGGEN_CLKSEL_BIT_L);
 
     bitFieldVal = (uint32_t)clkGenOutput <<
-                  (sigGenNum * EPG_GCTL1_SIGGEN_CLKSEL_BIT_L);
+                  ((uint32_t)sigGenNum * EPG_GCTL1_SIGGEN_CLKSEL_BIT_L);
 
     regValue = HWREG(base + EPG_O_GCTL1) & ~bitFieldLoc;
     regValue |= bitFieldVal;
@@ -576,10 +578,11 @@ EPG_setClkGenPeriod(uint32_t base, EPG_CLKGEN clkGenNum,
     uint32_t regValue, regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(clkGenNum <= 1U);
+    ASSERT((uint16_t)clkGenNum <= 1U);
     ASSERT(periodValue <= EPG_CLKDIV0_CTL0_PRD_M);
 
-    regLoc = EPG_O_CLKDIV0_CTL0 + (clkGenNum * EPG_CLKDIV_REG_OFF);
+    regLoc = (uint32_t)EPG_O_CLKDIV0_CTL0 + ((uint32_t)clkGenNum *
+                                             EPG_CLKDIV_REG_OFF);
 
     regValue = HWREG(base + regLoc) & ~(uint32_t)EPG_CLKDIV0_CTL0_PRD_M;
     regValue |= periodValue;
@@ -613,9 +616,10 @@ EPG_setClkGenStopEdge(uint32_t base, EPG_CLKGEN clkGenNum,
     uint32_t regValue, regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(clkGenNum <= 1U);
+    ASSERT((uint16_t)clkGenNum <= 1U);
 
-    regLoc = EPG_O_CLKDIV0_CTL0 + (clkGenNum * EPG_CLKDIV_REG_OFF);
+    regLoc = (uint32_t)EPG_O_CLKDIV0_CTL0 + ((uint32_t)clkGenNum *
+                                             EPG_CLKDIV_REG_OFF);
 
     regValue = HWREG(base + regLoc) & ~(uint32_t)EPG_CLKDIV0_CTL0_CLKSTOP_M;
     regValue |= ((uint32_t)stopEdge << EPG_CLKDIV0_CTL0_CLKSTOP_S);
@@ -652,16 +656,17 @@ EPG_setClkGenOffset(uint32_t base, EPG_CLKGEN clkGenNum,
     uint32_t regValue, regLoc, bitFieldLoc, bitFieldVal;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(clkGenNum <= 1U);
+    ASSERT((uint16_t)clkGenNum <= 1U);
     ASSERT(clkGenOutputNum <= 3U);
     ASSERT(offsetValue <= EPG_CLKDIV0_CLKOFFSET_CLK0OFFSET_M);
 
-    bitFieldLoc = EPG_CLKDIV0_CLKOFFSET_CLK0OFFSET_M <<
+    bitFieldLoc = (uint32_t)EPG_CLKDIV0_CLKOFFSET_CLK0OFFSET_M <<
                   (clkGenOutputNum * EPG_CLKDIV_CLKOFFSET_BIT_L);
 
     bitFieldVal = offsetValue << (clkGenOutputNum * EPG_CLKDIV_CLKOFFSET_BIT_L);
 
-    regLoc = EPG_O_CLKDIV0_CLKOFFSET + (clkGenNum  * EPG_CLKDIV_REG_OFF);
+    regLoc = (uint32_t)EPG_O_CLKDIV0_CLKOFFSET + ((uint32_t)clkGenNum  *
+                                                  EPG_CLKDIV_REG_OFF);
 
     regValue = HWREG(base + regLoc) & ~bitFieldLoc;
     regValue |= bitFieldVal;
@@ -689,7 +694,7 @@ EPG_enableSignalGen(uint32_t base, EPG_SIGGEN sigGenNum)
     uint32_t regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint16_t)sigGenNum <= 0U);
 
     regLoc = EPG_O_SIGGEN0_CTL0 + (sigGenNum * EPG_SIGGEN_REG_OFF);
 
@@ -715,7 +720,7 @@ EPG_disableSignalGen(uint32_t base, EPG_SIGGEN sigGenNum)
     uint32_t regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint16_t)sigGenNum <= 0U);
 
     regLoc = EPG_O_SIGGEN0_CTL0 + (sigGenNum * EPG_SIGGEN_REG_OFF);
 
@@ -751,9 +756,10 @@ EPG_setSignalGenMode(uint32_t base, EPG_SIGGEN sigGenNum,
     uint32_t regValue, regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint16_t)sigGenNum <= 0U);
 
-    regLoc = EPG_O_SIGGEN0_CTL0 + (sigGenNum * EPG_SIGGEN_REG_OFF);
+    regLoc = (uint32_t)EPG_O_SIGGEN0_CTL0 + ((uint32_t)sigGenNum *
+                                             EPG_SIGGEN_REG_OFF);
 
     regValue = HWREG(base + regLoc) & ~(uint32_t)EPG_SIGGEN0_CTL0_MODE_M;
     regValue |= (uint32_t)sigGenMode << EPG_SIGGEN0_CTL0_MODE_S;
@@ -782,9 +788,10 @@ EPG_enableBitRevOnDataIn(uint32_t base, EPG_SIGGEN sigGenNum)
     uint32_t regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint16_t)sigGenNum <= 0U);
 
-    regLoc = EPG_O_SIGGEN0_CTL0 + (sigGenNum * EPG_SIGGEN_REG_OFF);
+    regLoc = (uint32_t)EPG_O_SIGGEN0_CTL0 + ((uint32_t)sigGenNum *
+                                             EPG_SIGGEN_REG_OFF);
 
     HWREG(base + regLoc) |= EPG_SIGGEN0_CTL0_BRIN;
 }
@@ -810,9 +817,10 @@ EPG_disableBitRevOnDataIn(uint32_t base, EPG_SIGGEN sigGenNum)
     uint32_t regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint16_t)sigGenNum <= 0U);
 
-    regLoc = EPG_O_SIGGEN0_CTL0 + (sigGenNum * EPG_SIGGEN_REG_OFF);
+    regLoc = (uint32_t)EPG_O_SIGGEN0_CTL0 + ((uint32_t)sigGenNum *
+                                             EPG_SIGGEN_REG_OFF);
 
     HWREG(base + regLoc) &= ~(uint32_t)EPG_SIGGEN0_CTL0_BRIN;
 }
@@ -837,9 +845,10 @@ EPG_enableBitRevOnDataOut(uint32_t base, EPG_SIGGEN sigGenNum)
 {
     uint32_t regLoc;
 
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint16_t)sigGenNum <= 0U);
 
-    regLoc = EPG_O_SIGGEN0_CTL0 + (sigGenNum * EPG_SIGGEN_REG_OFF);
+    regLoc = (uint32_t)EPG_O_SIGGEN0_CTL0 + ((uint32_t)sigGenNum *
+                                             EPG_SIGGEN_REG_OFF);
 
     HWREG(base + regLoc) |= EPG_SIGGEN0_CTL0_BROUT;
 }
@@ -865,9 +874,10 @@ EPG_disableBitRevOnDataOut(uint32_t base, EPG_SIGGEN sigGenNum)
     uint32_t regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint16_t)sigGenNum <= 0U);
 
-    regLoc = EPG_O_SIGGEN0_CTL0 + (sigGenNum * EPG_SIGGEN_REG_OFF);
+    regLoc = (uint32_t)EPG_O_SIGGEN0_CTL0 + ((uint32_t)sigGenNum *
+                                             EPG_SIGGEN_REG_OFF);
 
     HWREG(base + regLoc) &= ~(uint32_t)EPG_SIGGEN0_CTL0_BROUT;
 }
@@ -896,10 +906,11 @@ EPG_setDataBitLen(uint32_t base, EPG_SIGGEN sigGenNum, uint32_t bitLength)
     uint32_t regValue, regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint32_t)sigGenNum <= 0U);
     ASSERT(bitLength <= 0xFFU);
 
-    regLoc = EPG_O_SIGGEN0_CTL0 + (sigGenNum * EPG_SIGGEN_REG_OFF);
+    regLoc = (uint32_t)EPG_O_SIGGEN0_CTL0 + ((uint32_t)sigGenNum *
+                                             EPG_SIGGEN_REG_OFF);
 
     regValue = HWREG(base + regLoc) & ~(uint32_t)EPG_SIGGEN0_CTL0_BITLENGTH_M;
     regValue |= (bitLength << EPG_SIGGEN0_CTL0_BITLENGTH_S);
@@ -933,9 +944,10 @@ EPG_setData0In(uint32_t base, EPG_SIGGEN sigGenNum,
     uint32_t regValue, regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint32_t)sigGenNum <= 0U);
 
-    regLoc = EPG_O_SIGGEN0_CTL1 + (sigGenNum * EPG_SIGGEN_REG_OFF);
+    regLoc = (uint32_t)EPG_O_SIGGEN0_CTL1 + ((uint32_t)sigGenNum *
+                                             EPG_SIGGEN_REG_OFF);
 
     regValue = HWREG(base + regLoc) & ~(uint32_t)EPG_SIGGEN0_CTL1_DATA0_INSEL_M;
     regValue |= (uint32_t)data0Input;
@@ -969,9 +981,10 @@ EPG_setData63In(uint32_t base, EPG_SIGGEN sigGenNum,
     uint32_t regValue, regLoc;
 
     ASSERT(EPG_isBaseValid(base));
-    ASSERT(sigGenNum <= 0U);
+    ASSERT((uint32_t)sigGenNum <= 0U);
 
-    regLoc = EPG_O_SIGGEN0_CTL1 + (sigGenNum * EPG_SIGGEN_REG_OFF);
+    regLoc = (uint32_t)EPG_O_SIGGEN0_CTL1 + ((uint32_t)sigGenNum *
+                                             EPG_SIGGEN_REG_OFF);
 
     regValue = HWREG(base + regLoc) & ~(uint32_t)EPG_SIGGEN0_CTL1_DATA63_INSEL_M;
     regValue |= (uint32_t)data63Input << EPG_SIGGEN0_CTL1_DATA63_INSEL_S;
@@ -1001,7 +1014,8 @@ EPG_setData0Word(uint32_t base, EPG_SIGGEN sigGenNum, uint32_t data0)
 {
     ASSERT(EPG_isBaseValid(base));
 
-    HWREG(base + EPG_O_SIGGEN0_DATA0 + (sigGenNum * EPG_SIGGEN_REG_OFF)) = data0;
+    HWREG(base + EPG_O_SIGGEN0_DATA0 + ((uint32_t)sigGenNum *
+                                        EPG_SIGGEN_REG_OFF)) = data0;
 }
 
 //*****************************************************************************
@@ -1026,7 +1040,8 @@ EPG_setData1Word(uint32_t base, EPG_SIGGEN sigGenNum, uint32_t data1)
 {
     ASSERT(EPG_isBaseValid(base));
 
-    HWREG(base + EPG_O_SIGGEN0_DATA1 + (sigGenNum * EPG_SIGGEN_REG_OFF)) = data1;
+    HWREG(base + EPG_O_SIGGEN0_DATA1 + ((uint32_t)sigGenNum *
+                                        EPG_SIGGEN_REG_OFF)) = data1;
 }
 
 //*****************************************************************************
@@ -1050,7 +1065,7 @@ EPG_getData0ActiveReg(uint32_t base, EPG_SIGGEN sigGenNum)
     ASSERT(EPG_isBaseValid(base));
 
     return(HWREG(base + EPG_O_SIGGEN0_DATA0_ACTIVE +
-                 (sigGenNum * EPG_SIGGEN_REG_OFF)));
+                 ((uint32_t)sigGenNum * EPG_SIGGEN_REG_OFF)));
 }
 
 //*****************************************************************************
@@ -1074,7 +1089,7 @@ EPG_getData1ActiveReg(uint32_t base, EPG_SIGGEN sigGenNum)
     ASSERT(EPG_isBaseValid(base));
 
     return(HWREG(base + EPG_O_SIGGEN0_DATA1_ACTIVE +
-                 (sigGenNum * EPG_SIGGEN_REG_OFF)));
+                 ((uint32_t)sigGenNum * EPG_SIGGEN_REG_OFF)));
 }
 
 //*****************************************************************************
