+++ driverlib/f28004x/driverlib/clb.h
@@ -690,8 +690,8 @@ static inline void CLB_writeSWReleaseControl(uint32_t base,
 
     HWREG(base + CLB_LOGICCTL + CLB_O_GP_REG) =
             (HWREG(base + CLB_LOGICCTL + CLB_O_GP_REG) &
-                ~(0x1000000U << inID)) |
-                    (((uint32_t)val) << (24U + inID));
+                ~(0x1000000U << (uint32_t)inID)) |
+                    (((uint32_t)val) << (24U + (uint16_t)inID));
 }
 
 
@@ -716,8 +716,8 @@ static inline void CLB_writeSWGateControl(uint32_t base,
 
     HWREG(base + CLB_LOGICCTL + CLB_O_GP_REG) =
             (HWREG(base + CLB_LOGICCTL + CLB_O_GP_REG) &
-                ~(0x10000U << inID)) |
-                    (((uint32_t)val) << (16U + inID));
+                ~(0x10000U << (uint32_t)inID)) |
+                    (((uint32_t)val) << (16U + (uint16_t)inID));
 
 }
 
@@ -924,7 +924,8 @@ static inline void CLB_configGPInputMux(uint32_t base, CLB_Inputs inID,
 
     HWREGH(base + CLB_LOGICCTL + CLB_O_IN_MUX_SEL_0) =
         (HWREGH(base + CLB_LOGICCTL + CLB_O_IN_MUX_SEL_0) &
-         ~(CLB_IN_MUX_SEL_0_SEL_GP_IN_0 << inID)) | (gpMuxCfg << inID);
+         ~(CLB_IN_MUX_SEL_0_SEL_GP_IN_0 << (uint16_t)inID)) |
+        ((uint16_t)gpMuxCfg << (uint16_t)inID);
 }
 
 //*****************************************************************************
@@ -1439,7 +1440,7 @@ static inline void CLB_programHLCInstruction(uint32_t base,
                                              uint32_t instruction)
 {
     ASSERT(CLB_isBaseValid(base));
-    ASSERT(instructionNum < 32);
+    ASSERT(instructionNum < 32U);
 
     CLB_writeInterface(base, CLB_ADDR_HLC_BASE + instructionNum, instruction);
 }
