+++ driverlib/f2837xs/driverlib/clb.h
@@ -614,7 +614,8 @@ static inline void CLB_configGPInputMux(uint32_t base, CLB_Inputs inID,
 
     HWREGH(base + CLB_LOGICCTL + CLB_O_IN_MUX_SEL_0) =
         (HWREGH(base + CLB_LOGICCTL + CLB_O_IN_MUX_SEL_0) &
-         ~(CLB_IN_MUX_SEL_0_SEL_GP_IN_0 << inID)) | (gpMuxCfg << inID);
+         ~(CLB_IN_MUX_SEL_0_SEL_GP_IN_0 << (uint16_t)inID)) |
+        ((uint16_t)gpMuxCfg << (uint16_t)inID);
 }
 
 //*****************************************************************************
@@ -1129,7 +1130,7 @@ static inline void CLB_programHLCInstruction(uint32_t base,
                                              uint32_t instruction)
 {
     ASSERT(CLB_isBaseValid(base));
-    ASSERT(instructionNum < 32);
+    ASSERT(instructionNum < 32U);
 
     CLB_writeInterface(base, CLB_ADDR_HLC_BASE + instructionNum, instruction);
 }
